+++ docs/changelog.rst
@@ -9,6 +9,9 @@ Unreleased
 
 New features:
 
+- :doc:`plugins/lastgenre`: Add a ``--pretend`` option to preview genre changes
+  without storing or writing them.
+
 Bug fixes:
 
 For packagers: